Refactored CreateEventForm into its own component + renamed CreateEventForm page to CreateEventFormPage. Started setup for Confirmation component
