+++ frontend/src/App.js
@@ -4,7 +4,7 @@ import LandingPage from "./pages/LandingPage";
 import ViewEventsPage from "./pages/ViewEventsPage";
 import ReportPage from "./pages/ReportPage";
 import FAQPage from "./pages/FAQPage";
-import CreateEventForm from "./pages/CreateEventForm";
+import CreateEventFormPage from "./pages/CreateEventFormPage";
 
 // import GeoLocator from "./utils/geolocator";
 
@@ -17,7 +17,7 @@ function App() {
         <Route path="/view-events" component={ViewEventsPage} />
         <Route path="/report" component={ReportPage} />
         <Route path="/faq" component={FAQPage} />
-        <Route path="/create-event-form" component={CreateEventForm} />
+        <Route path="/create-event-form" component={CreateEventFormPage} />
       </Switch>
     </div>
   );
