+++ frontend/src/components/CreateEventForm/MapCreateEvent.js
@@ -1,13 +1,15 @@
 import React from "react";
+import { Button } from "reactstrap";
 
 import Map from "../Map/Map";
 
-const MapEventCreate = (props) => {
+const MapCreateEvent = ({ handleSubmitStepTwo }) => {
   return (
     <>
       <Map />
+      <Button onClick={() => handleSubmitStepTwo()}>Continue</Button>
     </>
   );
 };
 
-export default MapEventCreate;
+export default MapCreateEvent;
