+++ frontend/src/components/Map/Polygons.js
@@ -2,16 +2,16 @@ import React from "react";
 import { Polyline, Popup } from "react-leaflet";
 
 const Polygons = ({ polygons }) => {
-  console.log(polygons)
-	const mapped_polygons = polygons.map((polygon, index) => {
-		return (
-			<Polyline key={index} positions={[polygon.shape]}>
-				<Popup>{polygon.popup}</Popup>
-			</Polyline>
-		);
-	});
+  console.log(polygons);
+  const mapped_polygons = polygons.map((polygon, index) => {
+    return (
+      <Polyline key={index} positions={[polygon.shape]}>
+        <Popup>{polygon.popup}</Popup>
+      </Polyline>
+    );
+  });
 
-	return <>{mapped_polygons}</>;
+  return <>{mapped_polygons}</>;
 };
 
 export default Polygons;
